edit done

+++ src/components/ContactList/ContactList.js
@@ -1,15 +1,30 @@
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import { deleteContact, fetchContacts } from '../../redux/contactsOperations';
 import { useEffect } from 'react';
 import { getFilteredContacts } from '../../redux/selectors';
+import Modal from 'components/EditModal/EditModal';
+import { EditForm } from 'components/Forms/EditForm';
 
 export const ContactList = () => {
   const dispatch = useDispatch();
   const filteredContacts = useSelector(getFilteredContacts);
+  const [editModalOpen, setEditModalOpen] = useState(false);
+  const [editingContactId, setEditingContactId] = useState(null);
 
   const handleDelete = id => {
-    dispatch(deleteContact(id));
+    const shouldDelete = window.confirm(
+      'Are you sure you want to delete this contact?'
+    );
+    if (shouldDelete) {
+      dispatch(deleteContact(id));
+    }
+  };
+
+  const handleEdit = id => {
+    setEditModalOpen(true);
+    setEditingContactId(id);
   };
 
   useEffect(() => {
@@ -17,14 +32,28 @@ export const ContactList = () => {
   }, [dispatch]);
 
   return (
-    <ul>
-      {filteredContacts.map(contact => (
-        <li key={contact.id}>
-          {contact.name}: {contact.number}
-          <button onClick={() => handleDelete(contact.id)}>Delete</button>
-        </li>
-      ))}
-    </ul>
+    <div>
+      <ul>
+        {filteredContacts.map(contact => (
+          <li key={contact.id}>
+            {contact.name}: {contact.number}
+            <button onClick={() => handleDelete(contact.id)}>
+              Delete Contact
+            </button>
+            <button onClick={() => handleEdit(contact.id)}>Edit Contact</button>
+          </li>
+        ))}
+      </ul>
+      <Modal isOpen={editModalOpen} closeModal={() => setEditModalOpen(false)}>
+        <EditForm
+          originalContact={filteredContacts.find(
+            contact => contact.id === editingContactId
+          )}
+          contactId={editingContactId}
+          closeModal={() => setEditModalOpen(false)}
+        />
+      </Modal>
+    </div>
   );
 };
